Fix to #15763 - "collection selector was not NavigationExpansionExpression" when joining two FromSql()s
Problem was that navigation expansion wasn't recognizing FromSqlOnQueryable as something that needs to be visited, so the expression was not being converted to NavigationExpansionExpression.
SelectMany requires collection navigation to be a NavigationExpansionExpression, hence the error.
Fix is to add generic handling of simple queryable methods.
