+++ test/EFCore.Relational.Specification.Tests/Query/FromSqlQueryTestBase.cs
@@ -305,7 +305,7 @@ namespace Microsoft.EntityFrameworkCore.Query
             }
         }
 
-        [Fact(Skip = "#15763")]
+        [Fact]
         public virtual void FromSqlRaw_queryable_multiple_composed()
         {
             using (var context = CreateContext())
@@ -325,7 +325,7 @@ namespace Microsoft.EntityFrameworkCore.Query
             }
         }
 
-        [Fact(Skip = "Issue#15763")]
+        [Fact]
         public virtual void FromSqlRaw_queryable_multiple_composed_with_closure_parameters()
         {
             var startDate = new DateTime(1997, 1, 1);
@@ -351,7 +351,7 @@ namespace Microsoft.EntityFrameworkCore.Query
             }
         }
 
-        [Fact(Skip = "Issue#15763")]
+        [Fact]
         public virtual void FromSqlRaw_queryable_multiple_composed_with_parameters_and_closure_parameters()
         {
             var city = "London";
@@ -503,7 +503,7 @@ FROM [Customers]"))
             }
         }
 
-        [Fact(Skip = "Issue#15763")]
+        [Fact]
         public virtual void FromSqlInterpolated_queryable_multiple_composed_with_parameters_and_closure_parameters_interpolated()
         {
             var city = "London";
